+++ src/correct_rice_gauss.rs
@@ -4,14 +4,14 @@ use std::ops::*;
 use matlab_fun;
 use matrix_math;
 
-pub fn compute(snr : &DMat<f32>) -> DMat<f32> {
-	let coefs : [f32;9] = [-0.289549906258443, -0.0388922575606330, 0.409867108141953, -0.355237628488567,	0.149328280945610, -0.0357861117942093,	0.00497952893859122, -0.000374756374477592, 0.0000118020229140092];
+pub fn compute(snr : &DMat<f64>) -> DMat<f64> {
+	let coefs : [f64;9] = [-0.289549906258443, -0.0388922575606330, 0.409867108141953, -0.355237628488567,	0.149328280945610, -0.0357861117942093,	0.00497952893859122, -0.000374756374477592, 0.0000118020229140092];
 	
 	//Fc=Coefs(1)+Coefs(2).*a1+Coefs(3).*a1.^2+Coefs(4).*a1.^3+Coefs(5).*a1.^4+Coefs(6).*a1.^5+Coefs(7).*a1.^6+Coefs(8).*a1.^7+Coefs(9).*a1.^8;
 	
 	let rows_size = snr.nrows();
 	let cols_size = snr.ncols();
-	let mut fc : DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+	let mut fc : DMat<f64> = DMat::new_zeros(rows_size, cols_size);
 
    for r in 0..rows_size {
 	for c in 0..cols_size {
@@ -29,11 +29,11 @@ pub fn compute(snr : &DMat<f32>) -> DMat<f32> {
 	return fc_final;
 }
 
-fn is_less_then(mat: &DMat<f32>, value: f32) -> DMat<f32> {
+fn is_less_then(mat: &DMat<f64>, value: f64) -> DMat<f64> {
 	let rows = mat.nrows();
 	let cols = mat.ncols();
 
-	let mut result : DMat<f32> = DMat::new_zeros(rows, cols);
+	let mut result : DMat<f64> = DMat::new_zeros(rows, cols);
 
 	for r in 0..rows {
 		for c in 0..cols {
