+++ src/em_ml_rice2D.rs
@@ -6,13 +6,13 @@ use matrix_math;
 use filter2b;
 use approxl1_i0;
 
-pub fn compute(mat : &DMat<f32>, n : i32, ws : usize) -> (DMat<f32>, DMat<f32>){
+pub fn compute(mat : &DMat<f64>, n : i32, ws : usize) -> (DMat<f64>, DMat<f64>){
 	//jezeli błąd będzie za duży spróbować użyć f64
 	//szkoda że to wymaga przerobienia połowy programu...
 	//let mat64 = matlab_fun::cast_to_f64(&mat);
 	
-	let tt: f32 = (ws * ws) as f32;
-	let mask_value: f32 = 1.0/tt;
+	let tt: f64 = (ws * ws) as f64;
+	let mask_value: f64 = 1.0/tt;
 	let mask_size: usize = ws;
 	
 	let mat_to_2 = matrix_math::power_matrix_by_scalar(&mat, 2);
@@ -34,7 +34,7 @@ pub fn compute(mat : &DMat<f32>, n : i32, ws : usize) -> (DMat<f32>, DMat<f32>){
 	return (a_k, sigma_n);	
 }
 
-fn compute_a_k(mat_to_2_filter : &DMat<f32>, mat_to_4_filter: &DMat<f32>) -> DMat<f32> {
+fn compute_a_k(mat_to_2_filter : &DMat<f64>, mat_to_4_filter: &DMat<f64>) -> DMat<f64> {
 
 	let mat_to_2_filter_to_2 = matrix_math::power_matrix_by_scalar(&mat_to_2_filter, 2);
 	let a = mat_to_2_filter_to_2.mul(2.0);
@@ -48,7 +48,7 @@ fn compute_a_k(mat_to_2_filter : &DMat<f32>, mat_to_4_filter: &DMat<f32>) -> DMa
 	return matrix_math::matrix_sqrt(&max_b_sqrt);
 }
 
-fn compute_sigma_k(a_k: &DMat<f32>, mat_to_2_filter: &DMat<f32>) -> DMat<f32> {
+fn compute_sigma_k(a_k: &DMat<f64>, mat_to_2_filter: &DMat<f64>) -> DMat<f64> {
 	let a_k_to_2 = matrix_math::power_matrix_by_scalar(&a_k, 2);
 	let c = mat_to_2_filter.clone().sub(a_k_to_2);
 	
@@ -57,7 +57,7 @@ fn compute_sigma_k(a_k: &DMat<f32>, mat_to_2_filter: &DMat<f32>) -> DMat<f32> {
 	return d.mul(0.5);	
 }
 
-fn compute_next_a_k(a_k: &DMat<f32>, mat : &DMat<f32>, sigma_k: &DMat<f32>, mask_size: usize, mask_value: f32) -> DMat<f32> {	
+fn compute_next_a_k(a_k: &DMat<f64>, mat : &DMat<f64>, sigma_k: &DMat<f64>, mask_size: usize, mask_value: f64) -> DMat<f64> {	
 	let a_k_times_mat = matrix_math::mul_matrix_by_matrix_each_value(&a_k, &mat);
 	let ak_div_sigma = matrix_math::div_matrix_by_matrix_each_value(&a_k_times_mat, &sigma_k);
 	let approx = approxl1_i0::compute(&ak_div_sigma);	
@@ -68,7 +68,7 @@ fn compute_next_a_k(a_k: &DMat<f32>, mat : &DMat<f32>, sigma_k: &DMat<f32>, mask
 	return new_ak;
 }
 
-fn compute_next_signal_k(mat : &DMat<f32>, a_k: &DMat<f32>, mask_size: usize, mask_value: f32) -> DMat<f32> {
+fn compute_next_signal_k(mat : &DMat<f64>, a_k: &DMat<f64>, mask_size: usize, mask_value: f64) -> DMat<f64> {
 	let abs_mat = matlab_fun::abs(&mat);
 	let abs_to_2 = matrix_math::power_matrix_by_scalar(&abs_mat, 2);
 	let filtered = filter2b::filter_img(&abs_to_2, mask_size, mask_value);
