+++ src/filter2b.rs
@@ -3,7 +3,7 @@ use na::{DMat};
 use std::cmp::{min, max};
 use std::ops::*;
 
-pub fn filter_img(image_src : &na::DMat<f32>, mask_size : usize, mask_val : f32) -> na::DMat<f32> {
+pub fn filter_img(image_src : &na::DMat<f64>, mask_size : usize, mask_val : f64) -> na::DMat<f64> {
     let mut image = image_src.clone();
     assert_eq!(image.nrows(), image.ncols());
     let range = mask_size as i32 / 2;
@@ -15,11 +15,11 @@ pub fn filter_img(image_src : &na::DMat<f32>, mask_size : usize, mask_val : f32)
 
     for x in 0..size {
         for y in 0..size {
-            let mut acc = 0f32;
+            let mut acc = 0f64;
             for xd in -range..range+1 {
                 for yd in -range..range+1 {
-                    acc += image_src[(pos(x + xd), pos(y + yd))] as f32 *
-                        mask[(mask_pos(xd), mask_pos(yd))] as f32;
+                    acc += image_src[(pos(x + xd), pos(y + yd))] as f64 *
+                        mask[(mask_pos(xd), mask_pos(yd))] as f64;
                 }
             }
             image[(x as usize, y as usize)] = acc;
