+++ src/rice_homomorf_est.rs
@@ -8,14 +8,14 @@ use em_ml_rice2D;
 use lpf;
 use correct_rice_gauss;
 
-pub fn compute_for_uknown_snr(mat : &DMat<f32>, lpf: f32, modo: i32) -> (DMat<f32>, DMat<f32>){
+pub fn compute_for_uknown_snr(mat : &DMat<f64>, lpf: f64, modo: i32) -> (DMat<f64>, DMat<f64>){
 	let (m2, sigma_n) = em_ml_rice2D::compute(&mat,10, 3);
 	let snr = matrix_math::div_matrix_by_matrix_each_value(&m2, &sigma_n);
 	
 	return compute(&mat, &snr, lpf, modo);
 }
 
-pub fn compute(mat : &DMat<f32>, snr: &DMat<f32>, _lpf: f32, modo: i32) -> (DMat<f32>, DMat<f32>){
+pub fn compute(mat : &DMat<f64>, snr: &DMat<f64>, _lpf: f64, modo: i32) -> (DMat<f64>, DMat<f64>){
 	let (m2, sigma_n) = em_ml_rice2D::compute(&mat,10, 3);
 	
 	let m1= filter2b::filter_img(&mat, 5, 0.04);
@@ -31,7 +31,7 @@ pub fn compute(mat : &DMat<f32>, snr: &DMat<f32>, _lpf: f32, modo: i32) -> (DMat
 	let mapa2 = matrix_math::mat_map(&lpf2, |x| x.exp());
 	
 	//2./sqrt(2).*exp(-psi(1)./2) = 1.8874
-	let const_val: f32 = 1.8874;
+	let const_val: f64 = 1.8874;
 	let mapa_g = matrix_math::mat_map(&mapa2, |x| x * const_val);
 	
 	let mut rn_abs = DMat::new_ones(1,1);
@@ -62,10 +62,10 @@ pub fn compute(mat : &DMat<f32>, snr: &DMat<f32>, _lpf: f32, modo: i32) -> (DMat
 	return (mapa_r, mapa_g);
 }
 
-fn value_or_zeros(mat : &DMat<f32>, value: f32) -> DMat<f32> {
+fn value_or_zeros(mat : &DMat<f64>, value: f64) -> DMat<f64> {
    let rows_size = mat.nrows();
    let cols_size = mat.ncols();
-   let mut result: DMat<f32> = DMat::new_zeros(rows_size, cols_size);
+   let mut result: DMat<f64> = DMat::new_zeros(rows_size, cols_size);
  
    for r in 0..rows_size {
 	for c in 0..cols_size {
